requirements.txt
openai>=1.0.0
pytesseract
pdf2image
Pillow
python-dotenv
